// src/app/models/producto.ts
export class ProductoModel {
    prod_ccateg: string;
    prod_cdesc: string;
    prod_cmarca: string;
    prod_npeso: number;
    prod_nprev1: number;
    prod_nprev2: number;
    prod_nprev3: number;
    prod_nstock: number;
    umed_ccods: string;
    prod_activo: boolean;
    prod_cimg: string;
    prod_ccod: string;
    prod_cfile: string;
    constructor(
        
    ){    }
}